kernel: disable net daemon, fix memory errors, increase stack size

+++ moose/include/net/netdaemon.h
@@ -5,4 +5,4 @@
 
 int init_net_daemon(void);
 void free_net_daemon(void);
-int net_daemon_add_frame(void *frame, u16 size);
+int net_daemon_add_frame(void *frame, size_t size);
